==> 0_MIMIC_to_parquet.ipynb <==
import pandas as pd
import sys, os, utils
from tqdm import tqdm

if os.getcwd()[-4:] == "code":
    os.chdir('../')

icu  = "./data/mimic/mimic-iv-2.2/icu/"
hosp = "./data/mimic/mimic-iv-2.2/hosp/"
ed   = "./data/mimic/mimic-iv-ed-2.2/ed/"

icu_parquet  = "./data/mimic-iv-2.2-parquet/icu/"
hosp_parquet = "./data/mimic-iv-2.2-parquet/hosp/"
ed_parquet   = "./data/mimic-iv-2.2-parquet/ed/"

for i in tqdm(os.listdir(icu)):
    if not os.path.isfile(icu_parquet + i[:-7] + ".parquet"):
        print("Processing %s"%i)
        globals()["{}".format(i[:-7])] = pd.read_csv(icu + i,compression="gzip",low_memory=False)
        globals()["{}".format(i[:-7])] = utils.columns_to_datetime(globals()["{}".format(i[:-7])])
        globals()["{}".format(i[:-7])].to_parquet(icu_parquet + i[:-7] + ".parquet")
        print("Finish %s to parquet"%i)

for i in tqdm(os.listdir(hosp)):
    if not os.path.isfile(hosp_parquet + i[:-7] + ".parquet"):
        print("Processing %s"%i)
        globals()["{}".format(i[:-7])] = pd.read_csv(hosp + i,compression="gzip",low_memory=False)
        globals()["{}".format(i[:-7])] = utils.columns_to_datetime(globals()["{}".format(i[:-7])])
        globals()["{}".format(i[:-7])].to_parquet(hosp_parquet + i[:-7] + ".parquet")
        print("Finish %s to parquet"%i)

for i in tqdm(os.listdir(ed)):
    if not os.path.isfile(ed_parquet + i[:-7] + ".parquet"):
        print("Processing %s"%i)
        globals()["{}".format(i[:-7])] = pd.read_csv(ed + i,compression="gzip",low_memory=False)
        globals()["{}".format(i[:-7])] = utils.columns_to_datetime(globals()["{}".format(i[:-7])])
        globals()["{}".format(i[:-7])].to_parquet(ed_parquet + i[:-7] + ".parquet")
        print("Finish %s to parquet"%i)
